extract linkKeysToLocks into its own file
