+++ paywall/src/data-iframe/blockchainHandler/getLocks.js
@@ -1,5 +1,3 @@
-import { linkTransactionsToKey } from './keyStatus'
-
 /**
  * Retrieve lock information and construct default keys for the locks
  *
@@ -8,7 +6,7 @@ import { linkTransactionsToKey } from './keyStatus'
  * @param {web3Service} web3Service the web3Service, needed for getLock
  * @returns {{ locks, keys }}
  */
-export async function getLocks({ locksToRetrieve, web3Service }) {
+export default async function getLocks({ locksToRetrieve, web3Service }) {
   const newLocks = await Promise.all(
     locksToRetrieve.map(lockAddress => web3Service.getLock(lockAddress))
   )
@@ -17,32 +15,6 @@ export async function getLocks({ locksToRetrieve, web3Service }) {
     (allLocks, lock) => ({
       ...allLocks,
       [lock.address]: lock,
-    }),
-    {}
-  )
-}
-
-export async function linkKeysToLocks({
-  locks,
-  keys,
-  transactions,
-  requiredConfirmations,
-}) {
-  const lockArray = Object.values(locks)
-
-  // convert into a map indexed by lock address
-  // and link each key to its lock
-  return lockArray.reduce(
-    (allLocks, lock) => ({
-      ...allLocks,
-      [lock.address]: {
-        ...lock,
-        key: linkTransactionsToKey({
-          key: keys[lock.address],
-          transactions,
-          requiredConfirmations,
-        }),
-      },
     }),
     {}
   )
